Fix data type of return

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t i = 0;
+	unsigned int i = 0;
 
 	if (h == NULL)
 		return (0);
@@ -21,6 +21,7 @@ size_t print_list(const list_t *h)
 			printf("[0] (nil)\n");
 		else
 			printf("[%u] %s\n", h->len, h->str);
+
 		h = h->next;
 		i++;
 	}
